# perf: optimize PoseidonSponge::apply_s_box