--- a/console/network/environment/src/prelude.rs
+++ b/console/network/environment/src/prelude.rs
@@ -50,6 +50,7 @@ pub use core::{
         BitXor,
         BitXorAssign,
         Deref,
+        DerefMut,
         Div,
         DivAssign,
         Mul,