@Pratyush 2022-10-26
Thanks for catching this! We can actually avoid converting it to a field elment entirely; I do this in the latest commit. Locally on the hashing benchmarks, this gives me a 22-35% improvement.